requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pinecone-client>=3.0.0
openai>=1.12.0
ShopifyAPI>=12.5.0
google-cloud-pubsub>=2.19.0
google-cloud-scheduler>=2.13.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-json-logger>=2.0.7
httpx>=0.26.0
tenacity>=8.2.0
pytest>=8.0.0
